feature: refactor of how to get a repo info [009]
