+++ assets/script/_modules/Helpers/Helpers.js
